fix(webgl): make one_dimensional_texture_access more robust to floating-point issues
Under certain WebGL implementations, the existing implementation of one-dimensional texture access,
specifically the computation of the texture y-coordinate, was leading to a sufficient loss of
floating-point precision to result in the wrong texel being accessed in certain circumstances,
leading to display artifacts.  This change makes the floating point computations less susceptible to
this particular loss-of-precision issue.