feat(sliceview): directly support an additional affine coordinate transform
Previously, an affine coordinate transform could be applied to SliceView sources, but it resulted in
a new transformed ChunkSource, which made it more difficult to support a user interface for
dynamically changing the transform.  With this change, the coordinate transform is specified
directly as a property of the sliceview RenderLayer, and can change dynamically.