fix(sliceview): fix isValid logic for chunk layouts with multiple visible sources
Previously, if multiple sources used the same chunk layout but with different bounds on the chunk
grid, then isReady might never return true even once all of the data has been loaded, due to
erroneously waiting for out-of-bounds chunks to be available.

This issue is due to the optimization in SliceView of computing the set of visible chunks only on a
per-ChunkLayout basis, rather than for each visible source, using enlarged bounds on the chunk grid
sufficient for containing all visible sources using that chunk layout.

This change keeps that optimization but keeps track of the number of total chunks number of chunks
that are visible.