fix(layout): properly perist 'xy' layout state
Thanks to Nico Kemnitz for reporting and identifying this problem, and providing an initial fix.

Previously, a layout of "xy" was not properly serialized to the JSON state, due to a discrepancy in
the default layout for single-LayerGroupViewer layouts (4panel) and split LayerGroupViewer
layouts (xy).

This change fixes the bug by always serializing the layout state even if it is equal to the
"default" value.  The existing defaults for both the single LayerGroupViewer and split cases are
retained, however, for compatibility when restoring from existing links.

Additionally, this change replaces the prior behavior of defaulting to an xy layout when creating a
new LayerGroupViewer via drag-and-drop.  Instead, the layout of the source LayerGroupViewer is used.