Removes .ts extension from import statements.
The import statement isn't expecting the extension and doesn't require
it. It also breaks the code transpiled by tsc. Before the code has been
converted with tsc it will be found in a file with a .ts extension. and
afterwards it will be placed in a file with a .js extension to reflect
the changes to the source. However, the import statement is not altered
during the transpilation and will fail to find the file with the .ts
extension.