feat(chunk_manager): add support for sourceQueueLevel
Previously, there was no way for one ChunkSource to depend on chunks from another ChunkSource,
without the risk of deadlock due to filling up the download queue.  With this change, a new
sourceQueueLevel property of the ChunkSource is introduced.  There is a separate download queue for
each sourceQueueLevel; there is not any significant actual increase in number of concurrent
downloads because it is assumed that any request for ChunkSource with non-zero sourceQueueLevel is
blocking another download.  To ensure there are no deadlocks, if a given ChunkSource A depends on
chunks from a ChunkSource B, then A must have lower sourceQueueLevel than B.

This change also eliminates the unused COMPUTING state.  Computational sources are alraedy indicated
by the isComputational property.