feat(perspective_view): use larger receptive field for picking
Previously, picking was based on only the single pixel under the mouse cursor, which made it
difficult to pick thin objects.  For skeletons and line/bounding box annotations, as a workaround,
for the purposes of picking the line were rendered with greater thickness.  For meshes, however, no
such workaround existed.

This commit uses a variant of the approach implemented by William Silversmith in #112: picking is
based on a 23x23 pixel "receptive field".  The point closest to the center with picking information
is used.  The workaround previously employed for skeletons and annotations is no longer needed.