refactor(sliceview/volume/renderlayer): explicitly support parameterized shaders
This refactors the shader handling in sliceview volume RenderLayer to support
parameterized shaders (e.g. by the custom shader text and UI control
specifications in the case of image_renderlayer, presence of equivalences and
other maps in the case of segmentation_renderlayer) in a cleaner and more robust
way.

This also removes the requirement that all sources of a multiscale source use
the same chunk format.

Fixes #179.