feat: snap to center of voxel
In Neuroglancer, the "center" position vector is represented in units of voxels, and an integer `n`
corresponds to the boundary between two voxels, while `n + 0.5` corresponds to the center of voxel.

Previously, using the `snap` command, or typing an integer coordinate in the position entry widget,
set the position coordinate to the nearest integer, which corresponds to the boundary between two
voxels.

For cross section views centered on the boundary between two coordinates along the orthogonal ("z")
dimension, it is numerically unstable to determine which of the two planes to display, and this can
lead to confusing results where the plane switches when changing zoom level, or incrementing the "z"
coordinate by 1 does not change the plane that is displayed.

To avoid such problems, these boundary positions are best avoided, and it was, therefore,
unfortunate that the Neuroglancer UI encouraged such positions.

With this change, the `snap` command, the position entry widget, and the discrete movement commands
all now "snap" to a voxel center position.  This should greatly reduce the likelihood of confusing
voxel boundary behavior on cross section views.