feat: add support for Google Cloud Storage JSON API
Previously, the S3-compatible XML API was used.  With this change, the GCS JSON API is used by
default, but the XML API can still be used via a new "gs+xml" URL scheme.

The JSON API has the advantage of not requiring a CORS policy be set on the bucket, and also does
not have a default server-side cache policy of 1 hour; that cache policy, which must be overridden
on a per-object basis, often leads to stale data being served.