feat(perspective_view): enabling position picking from transparent pick-disabled layers
Previously, when the "pick" enable/disable option was added to segmentation layers, position-based
picking (i.e. right click) was still supported for pick-disabled non-transparent layers, as this is
useful for navigation.  However, position-based picking was not supported for pick-disabled
transparent layers, in order to allow picking "through" transparent layers; this behavior was
intended to support the case where a pick-disabled transparent layer corresponded to ROIs that may
contain segments (e.g. neurons) in a separate pick-enabled layer.

With this change, position-based picking is now also support for transparent pick-disabled layers,
but transparent pick-disabled layers have lower precedence than opaque or pick-enabled layers.  That
is, if a given mouse position corresponds to a ray that intersects a transparent pick-disabled
layer, but does not intersect any opaque or pick-enabled layers, then the position of the
transparent pick-disabled layer is returned.  Otherwise, the frontpost position intersecting an
opaque or pick-enabled layer is returned.