Now uses the <li> tag as a container for messages
The element attribute in the StatusMessage class now refers to a <div>
tag inside the <li> container, which holds the contents of the
notification. When destroying the notification, the container (outer <li> tag) will be removed.

This is a preparation to handle the new close button we are going to
add, without affecting the old scripts that appends buttons in the
notification.