Notifications moved to the top of the website
To avoid blocking the images, a new close button is added. Before, the
user is not allowed to close notifications, although some notifications
times out automatically and some will be destroyed after some actions
are done by the user.

Future work might be needed to disable the close option for some kinds
of messages to prevent loss of data, etc when the user closes the
notification box by accident.