Disabled the annotation scorllbar on x-axis #5
The newly added annotations may be blocked by the scrollbar, as
mentioned in the issue. Instead of reverting the order to show the new
annotations at top, which requires some fundamental change to the offset
functions for auto-focusing, we disable the scrollbar so that the last
annotation does not get blocked anymore.