Reworked the operations matrix #4
The bottom matrix (operations matrix) will now be applied on (instead of
override) the top (rotation) matrix. Speficially, we store a hidden
'base' top matrix, which will not change unless the top matrix is
manually changed (either by hand or by fetching a rotation matrix from
the server). Any operation on the second matrix will change the top
rotation matrix on display but will not change the hidden 'base' matrix.