Feature: Transform by explicit affine operations ActiveBrainAtlasPipeline#4
User can apply the affine operations explicitly by specifiying the
amount of translation, rotation, scaling.

Due to limitations, we store a 'hidden base matrix' which is not changed
unless the rotation matrix on top is manually changed.

When the affine operations are changed, the top transformation matrix
will also change so that the *final* (combined) transformation matrix is
displayed.

When the rotation matrix is not changed, changing the affine operations
will apply the transformation *on top* of the hidden base matrix. Any
change made on the affine operation panel will not change the base
matrix.