Feature: Fetch transformation matrix ActiveBrainAtlasPipeline#3
A button is added to allow users to fetch transformation matrix to fit
the brain to the atlas, according to the center of mass data from the
database.
The button will only show under the rotation matrices if the current tag
is a brain. We use the source location of the image layer to determine
which animal to query.
The dimensionTransform function is used for converting the 3D
transformation matrix to adapt to higher ranks (e.g. in the case when
we use 4-th dimension for channels, etc.). It is originally written by
Litao Qiao for a deprecated functionality.

Co-authored-by: Litao Qiao <checkpppp@gmail.com>