Enhancement: Frontend logic enhancement
Now users will be notified if they have not logged in. Users who are not
logged in will not be able to use the multi-user mode and the page will
load the database data only.
Added a log-in button which directs user to the log-in page, and direct
them back after a successful log-in.
Improved the display for user list. The list will not be shown if the
user is not logged in.
Seperated the log-out button from the list of users, only shown in the
multiuser mode.